refactor(preload): split preload.ts into thin entry + 5 bridge modules (#769)
SRP wave-2 PR-A. Mechanical extract of `electron/preload.ts` (473 LoC) into
a thin entry + 5 single-concern bridge files. Zero behavior change; the
`window.ccsm*` surface (keys, signatures, listener semantics) is byte-identical.

- electron/preload/index.ts                       — entry, calls 5 install*
- electron/preload/bridges/ccsmCore.ts            — db/i18n/version/scan/cwds/connection/updater
- electron/preload/bridges/ccsmPty.ts             — pty IPC + listener fan-out
- electron/preload/bridges/ccsmSession.ts         — state/title/activate/cwdRedirected + setActive/setName
- electron/preload/bridges/ccsmNotify.ts          — flash listener + markUserInput
- electron/preload/bridges/ccsmSessionTitles.ts   — get/rename/list/enqueue/flush
- electron/window/createWindow.ts                 — preload path now `preload/index.js`
